### 2023-11-09 #77 @ADevillers
[Re] Bootstrap Your Own Latent: A new approach to self-supervised learning
**Original article:** J.-B. Grill, F. Strub, F. Altché, C. Tallec, P. H. Richemond, E. Buchatskaya, C. Doersch, B. A. Pires, Z. D. Guo,
and M. G. Azar. “Bootstrap your own latent: A new approach to self-supervised learning.” In: arXiv preprint
arXiv:2006.07733 (2020)

**Scientific domain:** Representation Learning
**Suggested editor:** @rougier
