### 2023-11-22 #77 @benoit-girard
@ADevillers : given the fact that you results were obtained using HPC, that the reviewers may not have access to, are there ways to test you code without it (like running only a subpart of it...)?

(same question applies to the companion paper)
