### 2023-11-23 #77 @ADevillers
Yes, you can train on CIFAR10 or perform linear evaluation on CIFAR10 and ImageNet with the following setup:

1. Ensure you use `--computer='other'`.
2. For GPU usage (recommended for faster processing), set `--hardware='mono-gpu'`. If you're using a CPU, set `--hardware='cpu'`.
3. Follow the commands provided in the README's final section.

Note: Training on ImageNet without HPC is impractical due to the large memory requirements for the necessary batch size and the extended duration it might take; yet, evaluation may be fine.
