### 2023-12-04 #77 @benoit-girard
Good news: @charlypg has accepted to review this paper and its companion!
